Move watch functionality over to CopyFilesPlugin
